+++ EXAMPLES/discriminant_analysis/discriminant_analysis_4.cpp
@@ -121,8 +121,8 @@ void discriminant_analysis_4_example () {
     XEM::LearnModelOutput * lMOutput = NULL;
 
     // TODO: next 3 lines should not be needed (always use vector ?)
-    int64_t* vLabel_p = lInput->getKnownLabelDescription()->getLabel()->getTabLabel();
-    vector<int64_t> vLabel;
+    int* vLabel_p = lInput->getKnownLabelDescription()->getLabel()->getTabLabel();
+    vector<int> vLabel;
     for (int j=0; j<lInput->getNbSample(); j++) vLabel.push_back(vLabel_p[j]);
 
     if ( lOutput->atLeastOneEstimationNoError() ) {
@@ -157,7 +157,7 @@ void discriminant_analysis_4_example () {
         //lMOutput->getCVLabel()->getLabel()->edit(std::cout);
         std::cout << "CV Error Rate:" << ( 1-lMOutput->getCVLabel()->getLabel()->getErrorRate ( vLabel ) ) *100 << "%" << std::endl;
 
-        int64_t ** classif = lMOutput->getCVLabel()->getLabel()->getClassificationTab ( vLabel, lMOutput->getNbCluster());
+        int ** classif = lMOutput->getCVLabel()->getLabel()->getClassificationTab ( vLabel, lMOutput->getNbCluster());
         std::cout << "CV Classification Tab:" << std::endl;
         for ( int i=0; i<lMOutput->getNbCluster(); i++ ) {
             std::cout << " | " ;
